Write the right cost to the right product in recompute_cost_price
As some products can be skipped if they are services, the zip loop can be
shifted. So it is replaced by a dictionary which have also the advantage to
group the write of the same cost.
Also if all the products are services, the call to write is empty which is not
allowed.

issue5218
review22681002
(grafted from 796fc217791f01d89e77acee6bb4a0873f988e63)