Do not set effective_date on assign
When the effective date is set on assignation, if the move stays in this state
for more than 1 day, it is no more taken into account for stock quantity
computation.
But if the effective date is not set on assign, the stock quantity computation
must always include assigned moves when stock_assign is set with the exception
of delta computation which must always use the effective or planned date.

issue6371
review36261002