Fix shipment report when locations are the same or transit location is used
The inventory moves is empty when the locations of the shipment are the same.
The other moves field must be used.
When a transit location is used for internal shipment, the moves field to use
depend of the state. If it is already shipped then the incoming moves is used
to know where to store products and if it is not yet shipped than the outgoing
moves is used to pick the products.

issue7881
review50641002